Updated PHP to remove the need for the token.php file. Updated token generation to use the uniqid() instead of a random Int. Logic in login function is now: if the generated token can be assigned to the user who's email is ? and password is ? (? = inputted values), then they can be logged in. Also added in two new php files that support the functionalities of adding and retrieving location information from the MySQL database
Created a few new activities.
One to see all of the locations that users have added in a list form. This list can be searched for desired keywords through a searchbar at the top.
One allows the user to see more detailed information about each location on its own page. This is also where the user who created the location will be able to edit and/or delete the location from the database.
The last new activity allows the user to add a location to the database. The activity requests location permissions and then automatically fills out the state, town, latitude, and longitude values for the user. This way, other users will be able to see where exactly the spot is, and users will not be able to create spots that they have not actually been to. There is also space to name the location something and give it a brief description.